+++ src/main/java/nextstep/subway/line/utils/PathFinder/DijkstraCustomGraph.java
@@ -1,5 +1,6 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.line.utils.PathFinder;
 
+import nextstep.subway.line.domain.Line;
 import org.jgrapht.EdgeFactory;
 import org.jgrapht.graph.WeightedMultigraph;
