+++ src/main/java/nextstep/subway/line/utils/PathFinder/DijkstraWeightedEdgeWithLine.java
@@ -1,5 +1,6 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.line.utils.PathFinder;
 
+import nextstep.subway.line.domain.Line;
 import org.jgrapht.graph.DefaultWeightedEdge;
 
 public class DijkstraWeightedEdgeWithLine extends DefaultWeightedEdge {
