+++ src/main/java/nextstep/subway/line/utils/PathFinder/PathFinder.java
@@ -1,9 +1,11 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.line.utils.PathFinder;
 
 import static org.springframework.util.ObjectUtils.isEmpty;
 
 import java.util.List;
 import nextstep.subway.exception.PathFindException;
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.line.domain.Section;
 import nextstep.subway.station.domain.Station;
 import org.jgrapht.GraphPath;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
